feature: 移植迫害生草图生成器插件；
